+++ Cadmus.Seed.Itinera.Parts/Codicology/MsCompositionPartSeeder.cs
@@ -47,9 +47,9 @@ namespace Cadmus.Seed.Itinera.Parts.Codicology
                         f => f.PickRandom("paper", "parchment"))
                     .RuleFor(s => s.Location, f => new MsLocation
                     {
-                        N = n,
+                        N = (short)n,
                         V = n % 2 == 0,
-                        L = f.Random.Number(1, 20)
+                        L = (short)(f.Random.Number(1, 20))
                     })
                     .RuleFor(s => s.Date,
                         HistoricalDate.Parse(("1300" + n)
@@ -67,15 +67,15 @@ namespace Cadmus.Seed.Itinera.Parts.Codicology
                     .RuleFor(s => s.Label, f => f.Lorem.Sentence(1, 3))
                     .RuleFor(s => s.Start, f => new MsLocation
                     {
-                        N = (n - 1) * 2,
+                        N = (short)((n - 1) * 2),
                         V = n % 2 == 0,
-                        L = f.Random.Number(1, 20)
+                        L = (short)(f.Random.Number(1, 20))
                     })
                     .RuleFor(s => s.End, f => new MsLocation
                     {
-                        N = (n - 1) * 2 + 1,
+                        N = (short)((n - 1) * 2 + 1),
                         V = n % 2 == 0,
-                        L = f.Random.Number(1, 20)
+                        L = (short)(f.Random.Number(1, 20))
                     })
                     .RuleFor(s => s.Date,
                         HistoricalDate.Parse(("1300" + n)
